campaign root updated with ts and bulk relation dialog container
